pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jarvis"
version = "0.1.0"
description = "JARVIS — an AI research agent powered by Claude, inspired by Iron Man"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "tavily-python>=0.3.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "chromadb>=0.5",
    "playwright>=1.44",
    # Production API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "prometheus-client>=0.20",
    "structlog>=24.0",
    "pika>=1.3",
    "apscheduler>=3.10",
    "sqlalchemy>=2.0",
    # OpenTelemetry
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    # Auth
    "pyjwt>=2.8",
    # Edge agent
    "paho-mqtt>=1.6",
]

[project.optional-dependencies]
voice = [
    "openai-whisper>=20231117",
    "pyaudio>=0.2.13",
    "numpy>=1.24",
    "pyttsx3>=2.90",
]
wake-word = [
    "pvporcupine>=3.0",
    "pyaudio>=0.2.13",
]
voice-cloud = [
    "openai-whisper>=20231117",
    "pyaudio>=0.2.13",
    "numpy>=1.24",
    "elevenlabs>=1.0",
]
vision = [
    "opencv-python>=4.9",
    "ultralytics>=8.0",
]
dashboard = [
    "streamlit>=1.35",
    "requests>=2.31",
]
docs = [
    "pypdf>=4.0",
    "python-docx>=1.1",
]

[project.scripts]
jarvis = "jarvis.main:main"
jarvis-api = "jarvis.api.server:main"
jarvis-worker = "jarvis.queue.worker:main"
jarvis-edge = "jarvis.edge.main:main"
jarvis-eval = "jarvis.evals.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
